Move dc_arg to dataclass module
In prep for adding attr_arg module
